+++ src/data-fetching/defineVueFireLoader.ts
