+++ src/data-fetching/navigation-guard.ts
