+++ tests/data-loaders/loaders.ts
@@ -1,5 +1,5 @@
 import { vi } from 'vitest'
-import { defineBasicLoader } from '../../src/data-fetching_new/defineLoader'
+import { defineBasicLoader } from '../../src/data-fetching/defineLoader'
 
 export const dataOneSpy = vi.fn(async () => 'resolved 1')
 export const dataTwoSpy = vi.fn(async () => 'resolved 2')
